+++ v2g-liberty/rootfs/root/appdaemon/apps/v2g_liberty/octopus_price_data_manager.py
@@ -1,44 +1,45 @@
 """Module to read data from (publicly) available Britisch electricity price data"""
 
-from datetime import datetime
+from datetime import datetime, timedelta, timezone
+import pytz
 import json
 import isodate
-import requests
+import aiohttp
+import asyncio
+from aiohttp import ClientTimeout, ClientError
 import constants as c
 import log_wrapper
-from v2g_globals import time_round, get_local_now, convert_to_duration_string
+from v2g_globals import (
+    time_round,
+    get_local_now,
+    convert_to_duration_string,
+    is_local_now_between,
+)
+
 from appdaemon.plugins.hass.hassapi import Hass
 
 
 class ManageOctopusPriceData:
-    """
-    Class reads data from (publicly) available price data at:
-    https://api.octopus.energy/v1/products/
-
-    This code is based upon:
-    https://github.com/badguy99/octoblock
-
-    """
+    """Class reads data from (publicly) available price data"""
 
     COLLECTION_NAME: str = "results"
     PRICE_LABEL: str = "value_inc_vat"
     START_LABEL: str = "valid_from"
     END_LABEL: str = "valid_to"
 
-    CURRENCY: str = "GBP"
-    UOM: str = ""
+    UOM: str = "GBP/MWh"
     EMISSIONS_UOM: str = "kg/MWh"
-
-    # The data comes in pence/kWh and should be sent to FM in pound/MWh
-    # Example: 20.5 pence/kWh = 0,205 pound/kWh = 205 pound/MWh, factor is 10
-    PENCE_PER_KWH_TO_POUNDS_PER_MWH_FACTOR: int = 10
 
     daily_timer_id: str = ""
 
     fm_client_app: object = None
     get_fm_data_module: object = None
 
-    # Octopus price urls
+    # Octopus price urls:
+    # https://api.octopus.energy/v1/products/AGILE-24-04-03/electricity-tariffs/
+    # E-1R-AGILE-24-04-03-H/standard-unit-rates/
+    # ?period_from=2025-07-24T23:00Z&period_to=2025-07-25T23:00Z&page_size=1500
+
     BASE_URL: str = "https://api.octopus.energy/v1/products/"
     import_url: str = ""
     export_url: str = ""
@@ -46,14 +47,21 @@ class ManageOctopusPriceData:
     BASE_EMISSIONS_URL: str = "https://api.carbonintensity.org.uk/regional/intensity/"
     emission_region_slug: str = ""
 
-    first_try_time_get_data: str = ""
-    second_try_time_get_data: str = ""
+    # This module is specific to Octopus UK, all times will be UK times
+    UK_TZ = pytz.timezone("Europe/London")
+
+    # Octopus publishes price data around 16:00, add a little slack.
+    FIRST_TRY_TIME_GET_DATA: str = "16:15:14"
+
+    # If not successful retry every x minutes until this time (the next day)
+    CHECK_RESOLUTION_SECONDS: int = 30 * 60
+    TRY_UNTIL: str = "12:34:56"
 
     # For more details about the regions see:
     # https://energy-stats.uk/dno-region-codes-explained/
     # https://carbon-intensity.github.io/api-definitions/#region-list
     # A..P = DNO region letter, 1..14 = carbon_intensity region number
-    gb_dno_regions = {
+    GB_DNO_REGIONS = {
         "Eastern England": ("A", 10),
         "East Midlands": ("B", 9),
         "London": ("C", 13),
@@ -77,25 +85,29 @@ class ManageOctopusPriceData:
         self.__log = log_wrapper.get_class_method_logger(hass.log)
 
     async def initialize(self):
-        self.__log(f"Initializing ManageOctopusPriceData.")
+        self.__log("Initializing")
+
+        if c.TZ != self.UK_TZ:
+            self.__log(
+                f"Home Assistant timezone is not UK but {c.TZ}, V2G Liberty might not work correctly!",
+                level="WARNING",
+            )
+            # TODO: Warn user
 
         ##########################################################################
         # TESTDATA: Currency = EUR, should be GBP! Is set in class definition.
         ##########################################################################
-        # self.__log(f"initialize, TESTDATA: Currency = EUR, should be GBP!", level="WARNING")
-        # self.CURRENCY = "EUR"
+        # self.__log(
+        #     "- - - T E S T D A T A - - -: Currency = EUR, should be GBP!",
+        #     level="WARNING",
+        # )
+        # c.CURRENCY = "EUR"
+        # self.UOM = "EUR/MWh"
+        # c.TZ = self.UK_TZ
 
-        self.UOM = f"{self.CURRENCY}/MWh"
+        self.__log("Completed")
 
-        # Emission data is calculated based on the production data and is therefore available a little later
-        # than the price data.
-        self.first_try_time_get_data = "16:17:18"
-        self.second_try_time_get_data = "17:18:19"
-
-        await self.kick_off_octopus_price_management(initial=True)
-        self.__log(f"Completed Initializing ManageOctopusPriceData.")
-
-    async def kick_off_octopus_price_management(self, initial: bool = False):
+    async def kick_off_octopus_price_management(self):
         """
         'Second stage' of initialisation.
         To be called from 'initialize' and from the globals module collective_action()
@@ -106,15 +118,16 @@ class ManageOctopusPriceData:
         """
         if c.ELECTRICITY_PROVIDER != "gb_octopus_energy":
             self.__log(
-                f"Not kicking off ManageOctopusPriceData module. Electricity provider is not 'gb_octopus_energy'."
+                f"Not kicking off ManageOctopusPriceData module."
+                f"Electricity provider is not 'gb_octopus_energy' but {c.ELECTRICITY_PROVIDER}."
             )
             return
 
         un_initiated_values = ["unknown", "", "Please choose an option", None]
         if c.GB_DNO_REGION in un_initiated_values:
             self.__log(
-                f"Not kicking off ManageOctopusPriceData module. c.GB_DNO_REGION ('{c.GB_DNO_REGION}') is not "
-                f"populated (yet)."
+                f"Not kicking off ManageOctopusPriceData module."
+                f" c.GB_DNO_REGION ('{c.GB_DNO_REGION}') is not populated (yet)."
             )
             return
 
@@ -132,90 +145,86 @@ class ManageOctopusPriceData:
             )
             return
 
-        self.__log(f"kick_off_octopus_price_management starting!")
+        self.__log("starting")
 
-        region_char, region_index = self.gb_dno_regions.get(c.GB_DNO_REGION)
+        region_char, region_index = self.GB_DNO_REGIONS.get(c.GB_DNO_REGION)
 
         self.import_url = (
             f"{self.BASE_URL}{c.OCTOPUS_IMPORT_CODE}/electricity-tariffs/"
             f"E-1R-{c.OCTOPUS_IMPORT_CODE}-{region_char}/standard-unit-rates/"
         )
-        self.__log(
-            f"kick_off_octopus_price_management self.import_url: {self.import_url}."
-        )
+        self.__log(f"self.import_url: {self.import_url}.")
 
         self.export_url = (
             f"{self.BASE_URL}{c.OCTOPUS_EXPORT_CODE}/electricity-tariffs/E-1R-"
             f"{c.OCTOPUS_EXPORT_CODE}-{region_char}/standard-unit-rates/"
         )
-        self.__log(
-            f"kick_off_octopus_price_management self.export_url: {self.export_url}."
-        )
+        self.__log(f"self.export_url: {self.export_url}.")
 
         self.emission_region_slug = f"/fw48h/regionid/{region_index}"
 
         if self.hass.timer_running(self.daily_timer_id):
             await self.hass.cancel_timer(self.daily_timer_id, silent=True)
         self.daily_timer_id = self.hass.run_daily(
-            self.__daily_kickoff_prices_emissions, start=self.first_try_time_get_data
+            self.__daily_kickoff_prices_emissions, start=self.FIRST_TRY_TIME_GET_DATA
         )
 
-        # Always do the kickoff at startup.
-        await self.__daily_kickoff_prices_emissions()
+        # Always kickoff at startup, delay to give globals the time to get all settings right first.
+        await self.hass.run_in(self.__daily_kickoff_prices_emissions, delay=45)
 
-        self.__log(f"kick_off_octopus_price_management completed")
+        self.__log("completed")
 
     async def __daily_kickoff_prices_emissions(self, *args):
-        self.__log(f"__daily_kickoff_prices_emissions called")
+        self.__log("called")
         await self.__get_octopus_import_prices()
         await self.__get_octopus_export_prices()
         await self.__get_gb_region_emissions()
-        self.__log(f"__daily_kickoff_prices_emissions completed")
+        self.__log("completed")
 
     async def __get_octopus_import_prices(self, *args):
         """Get import (consumption) price data from Octopus and send then to FlexMeasures"""
-        self.__log(f"Called")
-        res = requests.get(self.import_url)
-        if res.status_code != 200:
-            self.__log(
-                f"__get_octopus_import_prices. Error {res.status_code}, res: {res.text}."
-            )
-            if self.hass.now_is_between(
-                self.first_try_time_get_data, self.second_try_time_get_data
-            ):
-                self.__log(
-                    f"__get_octopus_import_prices, retry once at {self.second_try_time_get_data}."
-                )
-                await self.hass.run_at(
+
+        async def try_rerun():
+            """Helper method to rerun this method with a delay"""
+            # Try the rerun until TRY_UNTIL, this has to stop somewhere...
+            if is_local_now_between(self.FIRST_TRY_TIME_GET_DATA, self.TRY_UNTIL):
+                await self.hass.run_in(
                     self.__get_octopus_import_prices,
-                    start=self.second_try_time_get_data,
+                    delay=self.CHECK_RESOLUTION_SECONDS,
                 )
-            return
 
-        try:
-            # The data structure might not be valid so a little extra care for reading values
-            prices = json.loads(res.text)
-            prices = prices[self.COLLECTION_NAME]
-            prices = list(reversed(prices))
-        except Exception as e:
-            self.__log(f"exception reading JSON: {e}.", level="WARNING")
-            return
+        res = await self.__request_data_aiohttp(
+            self.import_url, params=self.get_octopus_rates_url_params()
+        )
 
-        consumption_prices = []
-        for price in prices:
-            consumption_prices.append(
-                round(
-                    float(price[self.PRICE_LABEL])
-                    * self.PENCE_PER_KWH_TO_POUNDS_PER_MWH_FACTOR,
-                    2,
-                )
+        if res is None:
+            self.__log(
+                f"Get data failed, retry in {self.CHECK_RESOLUTION_SECONDS} sec."
             )
+            await try_rerun()
+            return
 
-        start = parse_to_rounded_local_datetime(prices[0][self.START_LABEL])
-        end = parse_to_rounded_local_datetime(prices[-1][self.END_LABEL])
+        prices = res.get(self.COLLECTION_NAME, None)
+        if prices is None:
+            self.__log(
+                f"Could not get '{self.COLLECTION_NAME}' from data '{res}', aborting.",
+                level="WARNING",
+            )
+            return
+
+        start = self._parse_to_rounded_uk_datetime(prices[-1][self.START_LABEL])
+        end = self._parse_to_rounded_uk_datetime(prices[0][self.END_LABEL])
         duration = int(float(((end - start).total_seconds() / 60)))
+        if duration < 1400:
+            # Very rare: less than a day of data retry
+            await try_rerun()
+            # Keep processing data that has been received so far.
         duration = convert_to_duration_string(duration)
 
+        consumption_prices = [
+            convert_price(price[self.PRICE_LABEL]) for price in reversed(prices)
+        ]
+        # self.__log(f"start: {start}, end: {end}, prices: {prices}.")
         if self.fm_client_app is not None:
             res = await self.fm_client_app.post_measurements(
                 sensor_id=c.FM_PRICE_CONSUMPTION_SENSOR_ID,
@@ -226,65 +235,68 @@ class ManageOctopusPriceData:
             )
         else:
             self.__log(
-                f"__get_octopus_import_prices. Could not call post_measurements on fm_client_app as it is None."
+                "Could not call post_measurements on fm_client_app as it is None."
             )
             res = False
 
-        self.__log(f"__get_octopus_import_prices res: {res}.")
+        self.__log(f"res: {res}.")
         if res:
             if self.get_fm_data_module is not None:
-                parameters = {"price_type": "consumption"}
-                await self.get_fm_data_module.get_prices(parameters)
+                # FM needs processing time for the just uploaded prices before they can be queried
+                self.hass.run_in(
+                    self.get_fm_data_module.get_prices_wrapper,
+                    delay=45,
+                    price_type="consumption",
+                )
             else:
                 self.__log(
-                    "__check_for_price_changes. Could not call get_consumption_prices on "
-                    "get_fm_data_module as it is None."
+                    "Could not call get_prices on get_fm_data_module as it is None."
                 )
 
     async def __get_octopus_export_prices(self, *args):
         """Get export (production) price data from Octopus and send then to FlexMeasures"""
-        self.__log(f"Called")
-        res = requests.get(self.export_url)
-        if res.status_code != 200:
-            self.__log(
-                f"__get_octopus_export_prices. Error getting export prices at {self.export_url}, "
-                f"status_code: {res.status_code}, res: {res.text}."
-            )
-            if self.hass.now_is_between(
-                self.first_try_time_get_data, self.second_try_time_get_data
-            ):
-                self.__log(
-                    f"__get_octopus_export_prices, retry once at {self.second_try_time_get_data}."
-                )
-                await self.hass.run_at(
+
+        async def try_rerun():
+            """Helper method to rerun this method with a delay"""
+            # Try the rerun until TRY_UNTIL, this has to stop somewhere...
+            if is_local_now_between(self.FIRST_TRY_TIME_GET_DATA, self.TRY_UNTIL):
+                await self.hass.run_in(
                     self.__get_octopus_export_prices,
-                    start=self.second_try_time_get_data,
+                    delay=self.CHECK_RESOLUTION_SECONDS,
                 )
-            return
 
-        try:
-            # The data structure might not be valid so a little extra care for reading values
-            prices = json.loads(res.text)
-            prices = prices[self.COLLECTION_NAME]
-            prices = list(reversed(prices))
-        except Exception as e:
-            self.__log(f"__get_octopus_export_prices. exception reading JSON: {e}.")
-            return
-
-        production_prices = []
-        for price in prices:
-            production_prices.append(
-                round(
-                    float(price[self.PRICE_LABEL])
-                    * self.PENCE_PER_KWH_TO_POUNDS_PER_MWH_FACTOR,
-                    2,
-                )
+        res = await self.__request_data_aiohttp(
+            self.export_url, params=self.get_octopus_rates_url_params()
+        )
+        if res is None:
+            self.__log(
+                f"Get data failed, retry in {self.CHECK_RESOLUTION_SECONDS} sec."
             )
+            await try_rerun()
+            return
 
-        start = parse_to_rounded_local_datetime(prices[0][self.START_LABEL])
-        end = parse_to_rounded_local_datetime(prices[-1][self.END_LABEL])
+        prices = res.get(self.COLLECTION_NAME, None)
+        if prices is None:
+            self.__log(
+                f"Could not get '{self.COLLECTION_NAME}' from data '{res}', aborting.",
+                level="WARNING",
+            )
+            return
+
+        start = self._parse_to_rounded_uk_datetime(prices[-1][self.START_LABEL])
+        end = self._parse_to_rounded_uk_datetime(prices[0][self.END_LABEL])
         duration = int(float(((end - start).total_seconds() / 60)))
+        if duration < 1400:
+            # Very rare: less than a day of data retry
+            await try_rerun()
+            # Keep processing data that has been received so far.
+
         duration = convert_to_duration_string(duration)
+
+        production_prices = [
+            convert_price(price[self.PRICE_LABEL]) for price in reversed(prices)
+        ]
+        # self.__log(f"start: {start}, end: {end}, prices: {prices}.")
 
         if self.fm_client_app is not None:
             res = await self.fm_client_app.post_measurements(
@@ -296,69 +308,68 @@ class ManageOctopusPriceData:
             )
         else:
             self.__log(
-                f"__get_octopus_import_prices. Could not call post_measurements on fm_client_app as it is None."
+                "Could not call post_measurements on fm_client_app as it is None."
             )
             res = False
 
-        self.__log(f"__get_octopus_export_prices res: {res}.")
+        self.__log(f"res: {res}.")
         if res:
             if self.get_fm_data_module is not None:
-                parameters = {"price_type": "production"}
-                await self.get_fm_data_module.get_prices(parameters)
+                # FM needs processing time for the just uploaded prices before they can be queried
+                self.hass.run_in(
+                    self.get_fm_data_module.get_prices_wrapper,
+                    delay=45,
+                    price_type="production",
+                )
             else:
                 self.__log(
-                    "__check_for_price_changes. Could not call get_production_prices on "
-                    "get_fm_data_module as it is None."
+                    "Could not call get_production_prices on get_fm_data_module as it is None."
                 )
 
     async def __get_gb_region_emissions(self, *args):
         """Get emission data for region and send then to FlexMeasures"""
-        self.__log(f"Called")
+        self.__log("called")
 
         now = str(get_local_now())
         start = now[:10] + "T00:00:00Z"
 
         emissions_url = f"{self.BASE_EMISSIONS_URL}{start}{self.emission_region_slug}"
-        self.__log(f"kick_off_octopus_price_management emissions_url: {emissions_url}.")
+        self.__log(f"emissions_url: {emissions_url}.")
 
-        res = requests.get(emissions_url)
-        if res.status_code != 200:
+        res = await self.__request_data_aiohttp(emissions_url)
+        if res is None:
+            if is_local_now_between(self.FIRST_TRY_TIME_GET_DATA, self.TRY_UNTIL):
+                await self.hass.run_in(
+                    self.__get_gb_region_emissions,
+                    delay=self.CHECK_RESOLUTION_SECONDS,
+                )
+            return
+
+        emissions = res.get("data", None)
+        if emissions is not None:
+            emissions = emissions.get("data", None)
+        if emissions is None:
             self.__log(
-                f"__get_gb_region_emissions. Error {res.status_code}, res: {res.text}."
+                f"Could not get 'data', 'data' from data '{res}', aborting.",
+                level="WARNING",
             )
-            if self.hass.now_is_between(
-                self.first_try_time_get_data, self.second_try_time_get_data
-            ):
-                self.__log(
-                    f"__get_gb_region_emissions, retry once at {self.second_try_time_get_data}."
-                )
-                await self.hass.run_at(
-                    self.__get_gb_region_emissions, start=self.second_try_time_get_data
-                )
             return
 
-        try:
-            # The data structure might not be valid so a little extra care for reading values
-            emissions = json.loads(res.text)
-            emissions = emissions["data"]["data"]
+        # Extracting the 'forecast' values
+        emission_intensities = [entry["intensity"]["forecast"] for entry in emissions]
 
-            # Extracting the 'forecast' values
-            emission_intensities = [
-                entry["intensity"]["forecast"] for entry in emissions
-            ]
-        except Exception as e:
-            self.__log(f"__get_gb_region_emissions. exception reading JSON: {e}.")
-            return
-
-        start = parse_to_rounded_local_datetime(emissions[0]["from"])
-        end = parse_to_rounded_local_datetime(emissions[-1]["to"])
+        start = self._parse_to_rounded_uk_datetime(emissions[0]["from"])
+        end = self._parse_to_rounded_uk_datetime(emissions[-1]["to"])
         duration = int(float(((end - start).total_seconds() / 60)))
         duration = convert_to_duration_string(duration)
 
         ##########################################################################
         # TESTDATA: EMISSIONS_UOM = "%", should be kg/MWh
         ##########################################################################
-        # self.__log(f"__get_gb_region_emissions, TESTDATA: self.EMISSIONS_UOM = %, moet kg/MWh zijn!", level="WARNING")
+        # self.__log(
+        #     " - - -  T E S T D A T A - - - : self.EMISSIONS_UOM = %, moet kg/MWh zijn!",
+        #     level="WARNING",
+        # )
         # self.EMISSIONS_UOM = "%"
 
         if self.fm_client_app is not None:
@@ -371,24 +382,100 @@ class ManageOctopusPriceData:
             )
         else:
             self.__log(
-                f"__get_octopus_import_prices. Could not call post_measurements on fm_client_app as it is None."
+                "Could not call post_measurements on fm_client_app as it is None."
             )
             res = False
 
-        self.__log(f"__get_gb_region_emissions res: {res}.")
+        self.__log(f"res: {res}.")
         if res:
             if self.get_fm_data_module is not None:
                 await self.get_fm_data_module.get_emission_intensities()
             else:
                 self.__log(
-                    "__get_gb_region_emissions. Could not call get_gb_region_emissions on "
-                    "get_fm_data_module as it is None."
+                    "Could not call get_gb_region_emissions on get_fm_data_module as it is None."
                 )
 
+    async def __request_data_aiohttp(self, url: str, params: dict = None):
+        timeout = ClientTimeout(total=13, connect=3, sock_read=10)
 
-# TODO: Make generic function in globals, it is copied from Amber module.
-def parse_to_rounded_local_datetime(date_time: str) -> datetime:
-    date_time = date_time.replace(" ", "T")
-    date_time = isodate.parse_datetime(date_time).astimezone(c.TZ)
-    date_time = time_round(date_time, c.EVENT_RESOLUTION)
-    return date_time
+        try:
+            async with aiohttp.ClientSession(timeout=timeout) as session:
+                async with session.get(url, params=params) as response:
+                    if response.status != 200:
+                        self.__log(
+                            f"Error {response.status} from url '{url}' with params '{params}'.",
+                            level="WARNING",
+                        )
+                        return None
+                    try:
+                        res = await response.json()
+                        return res
+                    except aiohttp.ContentTypeError:
+                        text = await response.text()
+                        return json.loads(text)
+                    except json.JSONDecodeError as e:
+                        self.__log(
+                            f"Exception reading JSON: {e} from url {url}.",
+                            level="WARNING",
+                        )
+                    except Exception as e:
+                        self.__log(
+                            f"Exception reading JSON: {e} from url {url}.",
+                            level="WARNING",
+                        )
+        except asyncio.TimeoutError:
+            self.__log(f"Timeout on url: {url}.", level="WARNING")
+        except aiohttp.ClientConnectorError:
+            self.__log(f"Connection error on url: {url}.", level="WARNING")
+        except ClientError as e:
+            self.__log(f"Client error: {e} on url: {url}.", level="WARNING")
+        except Exception as e:
+            self.__log(f"Unexpected error: {e} on url: {url}.", level="WARNING")
+
+    def get_octopus_rates_url_params(self):
+        """Get octopus specific rates url params:
+          - period_from: 23:00 totday (BTC) converted into UTC ISO format
+          - period_to: 23:00 tomorrow (BTC) converted into UTC ISO format
+          - page_size: 1500 to make sure all data needed gets sent
+
+        Today/tomorrow
+        """
+        now_uk = datetime.now(tz=self.UK_TZ)
+
+        if not is_local_now_between(self.FIRST_TRY_TIME_GET_DATA, "23:59:59"):
+            # Can happen when app is restarted
+            self.__log(
+                f"Now ({now_uk.strftime('%H:%M:%S')}) is between 00:00:00 and "
+                f"{self.FIRST_TRY_TIME_GET_DATA}. No new data for tomorrow yet, getting today's."
+            )
+            now_uk = now_uk - timedelta(days=1)
+
+        def uk_to_23h_utc_iso(dt_uk):
+            """Helper to set 23:00 UK time and convert to UTC ISO8601 Z string."""
+            dt_23_uk = dt_uk.replace(hour=23, minute=0, second=0, microsecond=0)
+            dt_utc = dt_23_uk.astimezone(timezone.utc)
+            return dt_utc.isoformat().replace("+00:00", "Z")
+
+        period_from = uk_to_23h_utc_iso(now_uk)
+        period_to = uk_to_23h_utc_iso(now_uk + timedelta(days=1))
+
+        params = {
+            "period_from": period_from,
+            "period_to": period_to,
+            "page_size": 1500,
+        }
+        self.__log(f"params = {params}")
+        return params
+
+    def _parse_to_rounded_uk_datetime(self, date_time: str) -> datetime:
+        date_time = date_time.replace(" ", "T")
+        date_time = isodate.parse_datetime(date_time).astimezone(self.UK_TZ)
+        date_time = time_round(date_time, c.EVENT_RESOLUTION)
+        return date_time
+
+
+def convert_price(price: float) -> float:
+    """Convert a price in pence per kWh to pounds per MWh"""
+    # The data comes in pence/kWh and should be sent to FM in pound/MWh
+    # Example: 20.5 pence/kWh = 0.205 pound/kWh = 205 pound/MWh, factor is 10
+    return round(float(price) * 10, 2)